// api/models/recommend-video-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Vilingo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface RecommendVideoResponse
 */
export interface RecommendVideoResponse {
    /**
     * 标识推荐的视频的唯一id
     * @type {number}
     * @memberof RecommendVideoResponse
     */
    'id': number;
    /**
     * 视频标题
     * @type {string}
     * @memberof RecommendVideoResponse
     */
    'title': string;
    /**
     * 视频封面的URL
     * @type {string}
     * @memberof RecommendVideoResponse
     */
    'cover': string;
    /**
     * 封面显示的高度
     * @type {number}
     * @memberof RecommendVideoResponse
     */
    'height': number;
    /**
     * 对应着当前推荐的视频的练习id
     * @type {number}
     * @memberof RecommendVideoResponse
     */
    'eid': number;
}