// api/models/recommend-video-fetch-response.ts
// tslint:disable
/**
 * Vilingo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { RecommendVideoResponse } from './recommend-video-response';

/**
 * 
 * @export
 * @interface RecommendVideoFetchResponse
 */
export interface RecommendVideoFetchResponse {
    /**
     * 推荐的视频
     * @type {Array<RecommendVideoResponse>}
     * @memberof RecommendVideoFetchResponse
     */
    data: Array<RecommendVideoResponse>;
    /**
     * 通知前端是否还有更多视频
     * @type {boolean}
     * @memberof RecommendVideoFetchResponse
     */
    hasNextPage: boolean;
    /**
     * 本次查询返回了多少条数据
     * @type {number}
     * @memberof RecommendVideoFetchResponse
     */
    count: number;
    /**
     * 返回指向当前最后一个视频的游标
     * @type {string}
     * @memberof RecommendVideoFetchResponse
     */
    nextCursor?: string;
}